# PERF: speed up replace() up to 4x via more explicit typing on input array